Move app config to correct package
