+++ app/src/main/java/com/duckduckgo/app/global/ViewModelFactory.kt
@@ -37,7 +37,7 @@ import com.duckduckgo.app.privacymonitor.ui.PrivacyPracticesViewModel
 import com.duckduckgo.app.privacymonitor.ui.ScorecardViewModel
 import com.duckduckgo.app.privacymonitor.ui.TrackerNetworksViewModel
 import com.duckduckgo.app.settings.SettingsViewModel
-import com.duckduckgo.app.settings.db.AppConfigurationDao
+import com.duckduckgo.app.global.db.AppConfigurationDao
 import com.duckduckgo.app.settings.db.SettingsDataStore
 import com.duckduckgo.app.trackerdetection.model.TrackerNetworks
 import javax.inject.Inject
